entr.jinja: truncate history timestamps to seconds
When history records are created using default current time for
the 'dt' field, it's provided at microsecond resolution.  The
fractional seconds are just noise when displaying entries so
in the entr.jinja template we truncate them to seconds resolution.
The min() filter in jinja.py was modified to allow for seconds
resolution in addition to the default minutes resolution.