Remove need for doubled quotes in csv files
Previously both the python csv-reading code and the Postgresql
code that loaded the the kw* tables from the kw*.csv files
used the conventional practice of quoting text that contained
delimiter or quote characters and doubling the embeded quotes.
We don't need to do that since we are using tabs as delimiters
which won't ever occur in the text (because we forbid that).
The csv reading code in jdb.Kwds.load_csv() was changed to not
use any quote characters.  The sql code in loadkw.sql was changed
to use backspace for a quote character since postgreql seems to
require something and we don't ever expect backspaces in the text.

The two csv files that had doubled quotes were updated.

Also made minor fixes to the doc and help text in xml-rt.sh.