Add csv headers, Kwds loads all columns
This is a cherry-pick of 200723-bc550a0 from branch 'ents'
with edits to remove the entity-specific changes.  We want
just the csv headers changes and the changes to jdb.Kwds to
load full records rather than just the id,kw,descr fields.

From 200723-bc550a0:
Previously jdb.Kwds.loadcsv() produced kw* records with only
the "id", "kw" and "descr" fields (unlike when intialized from
a database by Kwds.load() which provided all columns.)
To support the addition of entity information in the kw* data
when it is read directly from the csv files, the Kwds class
now supports reading all columns similarly to the database case.
This in turn requires that the csv files have header lines.
Which broke a lot of tests...
test_fmtxml in particular uses it own set of kw*.csv files.
They were all replaced with symlinks to the standard csv files
in jmdictdb/data/ except for kwgrp.csv which has some synthetic
data for testing.