Tests: add test for jmxml.py PoS fix in prev rev
New test checks that PoS tags are not propagated to a following
sense without any PoS.  This was a long present but unrecognised
bug fixed in 200723-bf2393f.  Verified that the test fails for
the expected reason in rev 200721-f9b68cb (before fix).

Several test_edparse tests now fail which will be addressed
later.