Tests: replace load-testdb.py with load-testdb.sh
Both are intended to accomplish the same task but the .py
script, since it used the jmdictdb api, checked the database
version and commonly failed when trying to load an old
version of the test database prepatory to upgrading it.
Since the actual process is pretty simple it was simplest
for the moment just to reimplement it in sh.