fmtxml.py: Use KW entity info, remove xmlkw import, global XKW
Remove the xmlkw import and global XKW which formerly provided
information about differences between the tag information in
jdb.KW and the entities used in the XML; that information is
now included directly in the jdb.KW items thanks to 200723-f0a1b60.
All tests pass including xml-rt on both jmdict and jmnedict.