fmtxml.py: restore entr() 'genhists' parameter, fix tests
This parameter enables the output of <info>, <audit>, <notes>
and <src_notes> elements in the output XML provided compat
is "jmex" (or None).

Rev 200723-bf2393f removed the code that effected the action
of the 'genhists' parameter (although not the parameter itself)
since it was only activated when compat="jmdicthist" and the
"jmdicthist" option was removed in that rev.

test_fmtxml.py:
- Added some tests for the 'compat' and 'genhists' parameters.
- Added some initialztion statements for XKW and some comments
  explaining their presence.