fmtxml.py: add "type" attribute to <entr_corp>
In the compat="jmex" mode an XML <entr_corp> element is
produced containing the name of the corpus the entry
belongs to.  This rev adds an attribute, 'type" to it
that gives the name of the type of the corpus (ie, one
of rows in kwsrc, by default: "jmdict", "jmnedict",
"examples", kanjidic".