jmxml.py: remove warning "lsource has ... no text"
Lines like the following in the XML being parsed:
  <lsource ls_wasei="y"/>
produced a warning message like:
  Seq 1031920: lsource has attribute(s)  but no text
but are common (several dozen ocurrances) in the current versions
of the EDRDG distributed JMdict XML file and seem to be normal.
Cosequently warning about them is unhelpful and the warning
generating code has been removed.