jmxml.py: finish implementation of make_enttab()
jmxml.make_enttab() implementation was incomplete.  The code
for the different default handling of keywords between "jmdict"
and "jmnedict" dtd types had not been finished.  Specifically
"jmdict" tags are valid by default in the absence of a
.ent["jmdict"] value enabling them; opposite is true for
jmnedict tags, they should be removed unless there is an
enabling .ent["jmnedict"] value for them.

Previous to this rev test_jmxml.Test_jmnedict.test_20_3001010
(which checks if the use of a MISC tag not in the jmnedict
in jmnedict XML generates a warning and is ignored) failed:
the invalid-for-jmnedict MISC tag was added to the test entry.