entrs2xml.py: generate separate open/close root tags.
The value of the --root option may now be suffixed with "+"
or "-" to indicate only the opening or closing root tag be
output.  This is to allow running entrs2xml.py multiple
times while appending to the same output file; the first
time with --root="+", the middle times with --no-root, and
the last time with --root="-".