entrs2xml.py: revive old "jmdict-ex" (now "jmex") xml format
Is now called "jmex" but purpose is same: to provide a way
to dump and restore any kind of entry on a JMdictDB database.
Have reversed my earlier thinking that this was unnecessary;
it is the only practical way to move entries from an existing
database into a newly created, clean database without bringing
a possibly mangled database schema along as well.

Most of the changes to implement this are in fmtxml.py.
Also revived the extended format DTD which was removed in
  200120-0c649f6: "entrs2xml.py: Remove multiple corpus capability"
This should not be as burdensome now that entities are auto-
generated (updates won't be needed for tag changes); it will
only need to be updated when structural changes are made to
the JMdict DTD which happens rarely.
The <corpus> element as not revived, code remains in place
but disabled.  Will review later.