fmtxml.py: fix failure in "jmex" mode on RINF "name" tag
When processing a recent copy of the EDRDG "jmdict" database
fmtxml.py failed with a KeyError when formatting the tags of
entry id#1056754 (seq=2609160) which is rejected but has an
RINF "name" (id=103) tag on it's 2nd reading.  fmtxml.entity()
erroneous tried to look the tag up as a jmdict tag which
failed ("name" is a kanjidic tag.)

Fixed and added test_fmtxml.Jmex.test_001 to confirm.