Tests: jmtest01.sql: deleted schema "imp"
The "imp" schema is created when a JMdictDB database is configured
and is used for importing bulk data.  It is not used currently for
any tests and constitutes unnecessary bloat.  Worse, it's an accident
waiting to happen since JMdictDB updates are not, under current
pratice, applied to its objects, only to the objects in schema
"public".