pgload.py: load a .pgi file into a JMdictDB database
This is a new script that consolidates the multiple SQL commands
used previously to load the .pgi file output by bin/jmparse.py
and friends (kdparse.py, exparse.py).  The import.sql script
that does the heavy lifting was updated to work with it as was
Makefile-db.

The main new feature pgload.py brings is the ability to work with
.pgi files containing multiple corpora.

It is incomplete in that it creates a new corpus for each imported
corpus; plans are to also allow "merging" imported entries into
existing corpora.

There was in the past another program that performed the same
function, python/jmload.py (deleted in rev 20180604-c90212c), but
that one did so by rewriting the .pgi file.  pgload.py continues
to use the import schema / sql load method that replaced jmload.py
but drives the sql from python code rather than requiring the
steps to be manually executed.

One difference from doing the sql steps manually as was done in
Makefile-db previous to this is that now the import schema is
dropped and recreated for each import rather than reused.