db: prohibit multiple A entries for a seq# (db-e99634)
Added a unique index on the "entr" table that prohibits any two
active (stat=2), approved (unap=False) (the combination is shown
as status "A" on the cgi entr.py page) entries from having the
same src (corpus) and seq (sequence number) values.

submit.py: rearranged the order of operations when an entry
is approved.  Formerly when a new approved entry was added to
the database, the new entry was added first, then the entry's
parent entry was deleted.  In between those two operations the
database transiently contained two active, approved entries with
the same corpus/seq# which would have violated the above contraint
(had it been present) and caused the submission to fail.  This
revision reverses the order of those two operations: delete the
parent entry, then add the new entry.  Since these operations are
done inside a transaction, we do not have to worry that an error
will leave the parent entry deleted but the new replacement entry
unadded.

test_submit.py: added three new tests to check the submission of
approved entries (there were none):
- Add new unapproved entry and approve.
- Edit an approved entry and approve.
- Add two approved entries, verify second add fails.
The first test is unaffected by this revision.  The second
passed before this revision was applied.  When the database
unique index was implemented the test failed as expected with:
  UniqueViolation: duplicate key value violates unique constraint "entr_src_seq_idx"
When the change to submit.py was implemented, it passed again.
Did not try the third test prior to applying the updates but it
passes now.