entr2xml.py: restore multi-corpus capability
The ability to write entries from more than one corpus
to the xml file was removed in rev 200120-0c649f6 but
has regained importance for writing multiple corpora to
a "jmex" formatted xml file -- needed to be able to dump
entries in a high-level format and restore to a new database
without carrying the database schema along which happens
with pg_dump.  Could do this before this rev but awkwardly:
run entrs2xml multiple times and concatenate the results.
Command changes:
-s/corpus option: now takes a comma separated list of corpora
  names or id numbers optionally prefixed with "!" to mean "all
  except those listed."
-s/--corpus was required, it is now optional again.
-D/--debug option replaced by -L/--loglevel.