tags.html,.asc: Set hist.userid in sql examples
In the example sql that adds history records to document entries
affected by a tag change, we now set hist.userid in addition to
hist.name.  The name field is not reliable as anyone can submit
entries with any name; the userid field can be set only by authorized
users or programs.