db: revise 035-8fac2c: remove control chars from hist.* data
Revise update to include updating hist.* fields to remove control
characters (mostly '\r' but a smattering of others too) that were
erroneously not filtered out of the data in the past.  Tab ('\t')
character are not removed at this time.

Added a couple of very old (2011) updates to kwxref that escaped
being included in any updates up to now.

tests/xml-rt.sh: minor updates to option handling.  Unrelated
to the above change but not worth a separate commit.