db: revise 035-8fac2c: rename table kresolv to krslv
Tables xresolv and kresolv are referred to in code as xrslv
and krslv and intent has been to rename the tables themselves
to match.  This is harder to do for xresolv which is referenced
extensively but easier for kresolv since kanjdict support is
still not extensively used yet; will include in the 8fac2c
database update.

Also fix syntax eror in db/import.sql due to a premature
and incomplete edit that was accidently commited.

Generated an updated jmtest01 database by applying the
revised 035-8fac2c update to jmtest01.sql extracted from rev
200807-cc2412d8 which was the most recent one at db-c3ee8f
immediately prior to db-e99634 being applied to it.