jmapp: srchform, srchres and entr pages now work
Fixed imports in views/{srchform,srchres,entr}.py

In the view handlers for each of the above in jmapp.py, added
a "sid" item to the data passed to the jinja templates.  This
allows using the same templates as the cgi scripts use; the
jmapp-specfic templates in jmapp/tmpl/ were replaced with
symlinks to the cgi ones except for srchres.jinja which seems
to have other changes.

jdb.py: fix typo made when manually merging master into jmapp.