jmapp: fix imports in views; most pages now work
Also attempted to replace the jmapp Jinja templates with ones
from the cgi code.  This required changing the view functions
in jmapp.py to add a dummy 'sid' variable to the data passed
to the jinja templates.  This worked for several templates but
not others.  The ones it worked for are now symlinks to the
corresponding cgi template files.  edform.jinja and user.jinja
required the cgi template to be slightly modified to define the
'dbg' variable if it wasn't.  The cgi version of layout.jinja
lost css so was reverted.

All the pages seem to work except edconf.py (and possibly
submit.py; can't test without going through edconf) which
fails with an exception related to the recent serialization
changes.