cgi: Fix search by "nf" freq values
Previously, in srchform.jinja's Frequency of Use section, a "nf"
criterion would be added to the search if there was a value in the
"nfval" text box.  The "nf" checkbox was ignored.

The "nf" checkbox was removed and inclusion now determined by a
value in the "nfcmp" selection box.  Added a blank entry so user
can remove the criterion.  Also moved the P checkbox to the front
of the row (was at end.)

Validate the values of 'nfcmp' and 'nfval' to avoid potential
sql injection.

Same changes were made in srch.py used by jmapp.