Tests: runtests.py: revert the "t" prefix change
In revision
  200903-8d2ca19 tests: modules named T* and methods named t* are now tests
runtests.py was changed to consider any testcase method starting
with "t" to be a test allowing shorter test names than the default
"test" prefix allows.  Unfortunately the test methods share the
 same namespace as all the operational methods (setUp(), tearDown(),
skipTest(), all the assert* methods, etc) and runtests.py was
running tearDown() as a test.  This was most obvious when running
runtests.py with the "-v2" option which produced messages like:
  tearDownClass (tests.test_fmtjel.Restr)
  Hook method for deconstructing the class fixture after running all tests in the class. ... ok

Since there are no tests using the short prefi yet it was reverted
back to "test".