entrs2xml.py: fix crash when jmnedict entries have "abbr" tag
The JMdictDB database design allows entries belonging to any corpus
to have any tag; nor does the application code limit what tags can
be applied to entries in a given corpus yet (see IS#190).
But the XML generation code, recently revised to create the XML DTD
dynamically, would crash with a KeyError if it encountered a tag
that is invalid for that entry's corpus.  Turned out the the EDRDG
database contained several jmnedict entries that had been given the
tag "abbr" which is valid only for jmdict entries.  This resulted in
the new entrs2xml.py script crashing.

This update fixes that problem: the bogus tag will be ignored and
a warning message will be printed.  This (sans the warning message)
was the behavior prior to the dynamic DTD changes.