jdb.Kwds.upd(): add 'ents' parameter, allow None-valued args
Formerly the jdb.Kwds.upd() method used None for the default
value of the kw, descr parameters and did not provide an 'ents'
parameter at all.  The use of None for a default parameter value
precluded using an argument value of None to set a record's
'descr' or 'ents' field to None.

jmxml.py: adjusted for the change in jdb.Kwds.upd() parameter
defaults.
jdb.py: removed usused crossprod() function.