tests: jmdb.py: fix problem with testsrc table ownership
Previously when jmdb.py reloads a test database it droped and
recreated the table ("testsrc") used to hold the loaded file's
hash value.  When this happens the testsrc table is owned by
whomever happened to run the tests that caused the reload and
a subsequent reload by a different user will fail to drop the
table because it is owned by someone else.
In git-200903-b355c6c an attempt was made to fix this by recreating
the table with the standard JMdictDB user, "jmdictdb", but that too
requires special permissions.

Rather than dropping and recreating the table we now leave it in
place and just delete and recreate the single record in it which
anyone can do without any special permission and regardless of who
the table is owned by.

Also reorganised the code a little by promoting methods to functions
when they didn't use any instance variables.