tools: add kwcheck.py to compare database and csv file tags
This program compares the tags in a database's kw* tables with
a set of kw* CSV files.  Differences are reported in the form
of sql statements that would bring the database into conformance
with the CSV files.

db.py: add option to required() to return the updates present
 in the database in addition to those that are missing.  This
 information is used for reporting by kwcheck.py.
jdb.py: Kwds.loaddb(), .loadcsv(): Set new Kwds.source attribute
 with a name identifying the source the keyword data was loaded
 from.  For databases this is the database name (should be full
 uri); for csv files it is the directory path.  This information
 is used for reporting by kwcheck.py.