tests: test_fmtxml.Tags: remove kwsrc from setUp()
Previously the Tags testcase used the setUp() method to add a
"jmnedict" corpus to the test kwsrc table so the entry being
tested could reference it.  That turned out to be unnecessary
since the corpus type is explicitly specified (in the do_entr()
call) when the test entry is formatted.  Removed kwsrc setup
to avoid giving false impression it is needed.