jdb.setkeys(): Allow setting FK and order fields to None
In the database, the table rows that constitute a single
entry are linked together via foreign keys (e.g. kanj.entr
contains the entr.id value of the entry to which the kanj
row belongs) and when multiple rows are linked to the same
parent object they are ordered by a field typically named
"ord" or named the same as the table name (e.g. kanj.kanj).
In the Python code the objects that constitute an Entr object
are images of the corresponding database rows and contain
the foreign key and ordering fields as in the database but
in Python these fields have no significance since rows are
linked to their parent object by virtue of being in a list
attached to the parent object, and ordering is determined
by their position in that list.

This revision causes jdb.setkeys(), which sets the values
of the foreign key and ordering fields in a Python Entr
instance, to set them all to None if the 'id' argument
to setkeys() is None.  This simplifies some things like
testing.