jmapp: cgiinfo.py view was generating 500 server error
Fixed fatal problem:
- In rev 200828-65d091d the function signature of config.cfgRead()
  was changed but cgiinfo.py was not updated to match.

Rewrote the cgiinfo.py view (which is completely different than the
cgi version) to print config info read from the .ini files.  In
particular [status] section that gives files that the config info
was read from is useful.

Removed ability (requirement?) to specify a config file location
in the URL parameters; it now shows info from the actual config
read when the app was started and what is used when processing
requests.