cgi: add ability to block by ip address
This extends the existing "status" system that blocks access for
maintenance or excessive load.  That system looks for a "flag" file
in the STATUS_DIR directory (defined in the config file) early in
the cgi processing and if found, redirects to a corresponding html
page.  In addition we now check for the existence of a file named
"status_blocked" and if it contains an ip address matching the CGI
variable REMOTE_ADDR, similarly redirects to a "you are blocked"
page.