jmapp: jmapp.py: Remove default for config file location.
We now require the environnment variable JMAPP_CFGFILE to be
defined and point to a readable config file at the time the
jmapp.py module is imported.  There is longer any default
value if JMAPP_CFGFILE is not defined (it was too hard to
figure out an appropriate value) and instead an exception is
raised.  An exception is also raised if the config file is
unreadable.  An operator can find the details in the web
server's log files.

Both bin/run-jmapp.py (for running the Flask debug server) and
web/cgi-bin/jmapp.wsgi (for running jmapp with Apache/mod_wsgi)
will set JMAPP_CFGFILE.