config.py: config filenames now relative to config.ini
Relative filenames in the configuration .ini file are now
resoloved relative to the location of the .ini file.  This
was previously the case for logging/LOG_FILENAME but not for
web/STATUS_DIR.  Consequently...

IMPORTANT: Any relative path in config.ini for [web]/STATUS_DIR
will now be resolved to a different file location and must be
changed of the former location is to be preserved.

config.py: Now converts relative file paths in the config.ini
file to absolute paths when constructing the config object that
holds all the config settings.  They are resolved relative to
the config.ini file's location.  Resolving them relative to the
current working directory was problematical since different servers
may have different cwd's.  Expanding them to full, absolute paths
in config.py removes abiguity about where the referenced file
actually is or should be when debugging.

cgi/cgiinfo.py, views/cgiinfo.py:
- Show location of logfile and status directory.
- Handle errors when expected config entries are missing.