jmapp: edconf.jinja was failing with undefined 'sid'
When the edconf page has warning messages, rendering of
edconf.jinja raised a jinja.UndefinedError exception on
variable 'sid' in block "edconf_warns".  The edconf.jinja
file is shared by both the cgi and wsgi (aka jmapp) code
and the method for handling the 'sid' values (which used
in the cgi code but not in the wsgi code) was applied to
all the jinja templates in rev 200828-6a593f7 but apparently
I missed this file.