jmapp: views edconf.py, edform.py: fix bad error handling
The views edconf.py and edform.py contained erroneous calls
to jmcgi.err_page() that were overlooked when porting from the
cgi code.  err_page() calls sys.exit() which is ok for the cgi
scripts but incorrect for wsgi (mod_wsgi prints a message to
the error log and ignores the call but the error page html ends
up in the log as well.  The wsgi views handle errors by returning
them to the caller and both scripts were fixed to do so.
Two other jmcgi functions call sys.exit(): check_server_status()
and redirect().  The wsgi code has iis own version of the former
in srvlib.py and the latter is not called from the wsgi code.