Install jmdictdb package with useful version number
When a jmdictdb package is built or installed now it is given a
version number derived from the current Git HEAD commit with the
format: YY.MM.DD.vxxxxxxx where YY, MM, DD are the year, month
and day of the commit and xxxxxxx is the commit hash.  The version
number is also available in code as a '__version__' attribute on
imported module 'jmdictdb'.

The version number is stored in the file jmdictdb.__version__.py
which is updated by tools/upd_version.py.  That program can be
run by hand when needed and/or setup as a Git post-commit hook
which will automatically update it after each commit.  More
details are in the comments in upd_version.py.

setup.py was updated to get the package version number from
__version__.py.  Also added package "flask" to the requirements.

Makefile now runs tools/upd_version.py before installing package
(we don't want to rely on the git hooks to keep __version__.py
updates.)

The cgi and jmapp versions of cgiinfo pages were updated to show
the jmdictdb package version number.