resetpw.py: new options to un-disable, make admin
The main purpose of this program is recourse for an Amin user
that accidently locks themself out of the system (forgets password,
changes account to 'disabled', removes Admin privilege, etc.  The
latter cases were not allowed for previously since the program
could only change the password.  This rev fixes that.