cgi: add "SameSite" cookie attribute, lengthen login timeout
In jmcgi.py added the attribute "SameSite=Strict" to the jmdictdb_sid
cookie to try to diagnose a premature logout problem reported by a
user.  This will be needed in future anyway as browsers are updated
to reject insecure cookies without this attribute.

Also lengthened the SESSION_TIMEOUT value from 2 hours to 4 and use
its value for both the database timeout and the sid cookie' max-age
attribute.

Added debug logging statements to the some of the db* functions.