jmsess: Fix timezone confusion (jmsess db-04f177)
Several of the db* functions in jmcgi.py run sql to compare a user's
session timestamp in the jmsess database sessions.ts column with the
current time returned by Postgresql function NOW().  The problem was
that the time in sessions.ts was set from a default value that was in
UTC but the time returned by NOW() is localtime resulting in bogus
comparisions if the server localtime timezone is not UTC.

Fixed by changing the sessions.ts default value to be NOW() rather
than the former expression that converted NOW() to UTC which required
update to mksess.sql for a new jmsess database and and new script
039s-04f177.sql to update an existing jmsess database.

rest.py: unrelated minor typo fix not worth a separate commit.