jmapp: srchres url parameter was ignored in srchres.py url
Normally the srchres.py (search results) page will redirect to the
entr.py page if the search resulted in finding only one entry.  However
it allows a &srchres=1 url parameter to override this behavior and display
the srchres page with the usual list of results (with only one item in it).
Processing of the parameter happens in jmapp.py but was overlooked when
the srchres view was migrated from the cgi version.  Fixed now.