html: html-escape kwfreq.descr text
Some rows in the kwfreq table contain quote (") characters in the
.descr field text and that text was making its way into entr.jinja
unescaped (via jinja.TALfreqs()) where it was used in the "title"
attribute of two elements, causing the attributes to render incorrectly.
The escaping (turning into '&quot;') needs to be done in jdb.freq2txt()
because it returns html which can't be additionally escaped later.
(Some grepping indicated that those were the only places the html
output of jdb.freq2txt() was used so nothing else should end up
double-escaped.)