cgi, jmapp: error reporting for jelparse syntax errors
Previously, jelparse syntax errors were reported by showing
the input JEL text with a with a line below containing a caret
(^) character under the approximate position of the error as
reported by parser.  This meant the error text had to use
"pre" formatting in the html to preserve whatspace.  But that
meant other message did not wrap and, if long, extended off
the right side of the window.  The was also a problem positioning
the caret correctly in presence of wider Japanese characters
despite specifying a mnonspace font.

We now insert a couple of blot (▇) characters into the input
JEL text at the error position reported by the parser.  This
allows the lines (and other error message lines) to be wrapped.

An outstanding problem of how to pass prolog and epilog text
to the error page from edconf.py view was also fixed.