jelparse.y: fix uncaught KeyError in mk_restrs()
When JEL input text was parsed that had a restriction (restr/stagr/stagk)
with an item that did not match anything in the target list (e.g. a kanji
word in a KR restr that did not exist in the entry's kanji) the call to
txt2restr() in mk_restrs() would raise a KeyError.  This was not caught
by anything and bubbled up to the top-level handler.  In the jmapp/WSGI
version it produced a stack dump to the JMdictDB log file and a generic
"something's wrong" error page which gave no clue about the source of the
problem.  This revision fixes it to be a ParseError handled like the other
conditions that raise that and with a useful error page message.

Also in jelparse.mk_restrs(), removed code that duplicated the target
lookup check that happens in restr.txt2restr() (which raises the exception
noted above).

Simplified the error message.  Since we report the restriction text that
caused the problem and the input text, it doesn't seem necessary to spell
out if it is kanji or reading, etc.

In the edconf.py CGI and WSGI pages, escape the error message prolog
text since it contains user-supplied data and is not excaped in the jinja
error page template.

Tests updated to account for the different exception raised by JELparse().