cgi: improved robustness of cgiinfo.py
The cgiinfo CGI page failed when a situation arose when it was unable
to import from the jmdictdb package.  It is intended that it should
still produce a page reporting the import problem but it instead crashed
due to a bug (wrong number of arguments when formating the 'Page' string
in the ImportError exception block.)  Took opportunity to reorganize and
improve the code a bit to make it present other info (like environment
variables) that are not dependent on the jmdictdb imports even if imports
fail.
Akso made parallel updates to the jmapp version in jmdictdb/views/ but
checking for import errors there is not very useful because it's called
by flaskapp.py which had to import the modules to run.