submit.py: Fixed return values in error case
submission() and some of the functions it calls normally return
a 3-tuple but when an error is detected, returned None.  This could
result in an exeption when a caller tried to unpack an expected
3-tuple return value.  submission() now returns (None,None,None)
in error cases.

test_submit.py was adjusted correspondingly.  Also added an additional
test case to check error messages when the parent of an entry
being edited disappearrs before it is submitted (as can happen when
another edit of the same entry is approved before the first edit is
submitted.)