flaskapp.py: svc parameter lost on srchres->entr redirect
When a search returned in a single result, the srchres.py route
handler in flaskapp.py failed to include the "svc" url parameter
when redirecting to the entr.py page.  If search was done in a
different svc/database than the default "jmdict", this resulted
in the wrong entry or an "entry not found" error being displayed
because the svc database would revert to "jmdict" rather than
the one the search was done in.