submit.py: rework database support for submissions (db-972bf3)
Replaced the database functions get_edroot and get_edtree with a
more general and versatile view, "edpaths", with the corresponding
(substantial) updates to jmdictdb/submit.py.  The edpath view
provides edit tree information in a form (a set of edit paths,
from root to each leaf entry) that is simpler to use in the
submit.py code.  These changes required a update of the database
schema version number.

The schema update also adds a new contraint to the "entr" table
that prohibits an approved entr from being an edit of some other
entry (ie, having a non-null .dfrm value).  This restriction was
maintained in the JMdictDB software previously.

tests/test_submit.py was updated to add more tests and new helper
functions were added to make adding tests easier.
jmtest01.sql: updated to the db-972bf3 schema.

The current tests pass with both the current and the previous
versions of submit.py.  To test with the previous version of
submit.py, replace the current version with the previous version
(extracted from Git) and manually add the get_edroot() and
get_subtree() database functions (copy from the previous version
of db/mkviews.sql and paste into psql) to the jmtest01 database;
the submit tests should still run and pass.

db/entrobjs.sql: in addition the the new entr_dfrm_check constraint,
trainling whitespace was deleted so the diff is a little noisy.