hist: Add column "eid" to "hist" table
This branch is to explore recording the entr.id value of the entry
item a hist item is first attached to, in the hist row itself.  The
hist.entr column references the entry the history item is currently
attached to which will be different when current item is the result
of edits and the original entry item may have been deleted.  Although
the orignal entry id may not reference any entry currently in the
database (thus can't be a foreign key) it is still useful since it
can be matched to entry id values recorded in log files and the log
file information can provide additional details about the addition
of the entry the history was first added to.

This revision adds column "eid" to table "hist" and makes the necessary
collateral changes elsewhere.  It does not yet set or otherwise use
that column.

The tests in tests/runtests.py all pass but no other testing has been
done yet.