hist: Set hist.eid in jdb.addentr()/.setkeys() (db-b5d00f)
[The db-b5d00f note should have been applied to 211225-3bfc8f1.]
New column ".eid" in table "hist" was added in revision 211225-3bfc8f1
and will hold the entry id of the entry the history item was first
added to.  That revision added the column but did not provide for
anything to set it.  This revision provides for automatically setting
a value in that column when an entry is submitted.

The hist.eid value is now set in jdb.addentr() (actually in
jdb.setkeys() but this is always called by jdb.addentr() before the
latter writes an entry to the database.  The .eid value is set only
on the last item in the _hist list which we presume is the hist item
for the current submission of the entry.  No provision is made for
using addentr() without making any changes to the last hist item.

Added tests for Entr._hist including the new .eid field in test_submit
and updated jmtest01 database to db-b5d00f.