jmsess: replaced resetpw.py with users.py
resetpw.py was previously used as a backup method of accessing
the jmess user database if it became inaccessible via the usual
access method of the users.py web page.  It had very limited
capabilities though, only allowing changes to passwords or privilge
level of existing users.

The new program, bin/users.py, allows full management of users
including listing, adding, deleting and modifying all their settings.
It is thus a superset of the capabilities of resetpw.py.
It also eliminates the need for the jmsess creation script to add
an initial admin user since that can now be done by bin/users.py.

The "list" action of users.py is currently a little crude but
will be improved in a future revision.