(ftom xng): Removed parse support for edict/edict2.
[Cherry picked from branch xng: 220116-34ec129]
Removed the Edict/Edict2 parser module (jmdictdb/edparse.py), its tests,
and it use in the Edit form (views/edform.py) where it could be used to
initialize the form fields.  It was never used AFAICT and it value-to-
maintenance-cost ratio was very low.
Edict/Ddict2 remains supported as an output format.