views/cgiinfo.py: add wsgi file location, db section
Added an additional item in the Execution section for the location of
the wsgi file.  To get that info requires adding an extra line of code
to the wsgi file (documented in a comment in cgiinfo.py).
Added a new section that reports the database and update id(s) for each
db_* service listed in the active config file.  The databases are
accessed with the credentials given in the config file.