tools/bulkupd.py: Add support for kinf/rinf tags
Required a means of specifying the particular kanji or reading
the tag change is to apply to.  Method used was to extend the
the syntax "[<sense#>]" to allow a kanji/reading ordinal number
or text in place of the sense# for kinf/rinf changes.

Also...
Catch bad sense number and handle gracefully rather than crashing.
Catch bad tag texts and handle gracefully rather than crashing.
Improved the wording of some error messages.
Fixed error causing "skipped" counts to be not reported in the
  total sumary results.
Added entry id number (in addition to seq#) to error messages.
Removed following unused Error objects:
 ChildError, MissingError, MultipleError, NoSenseError (p)
NotFoundError was replaced with NoItemError.