Makefile: dont install unused/missing programs
Don't try to install resetpw.py, it no longer exists (was removed in
211229-4eed89a.)  Its replacement, bin/users.py need not be installed
system-wide because it is for admin use only and is rarely needed as
it is used only during initial installation or as a backup for the
web Users page.

Don't install dbcheck.py. dbreaper.py, they are seldomly used and need
maintenance.