views/cgiinfo.py: Remove import error trapping
Earlier CGI version attempted to trap import errors of jmdictdb modules
and report them.  But that is ineffective with the wsgi version since
the jmdictdb Flask app has to be running to display the page and it
wouldn't be if any of the imports failed.  So simplfied code by ripping
that stuff out.

Also changed labels of some items to make clearer where their information
came from.