fmtxml.py: stop writing "illegal tag" warnings to stderr
Previously when fmtxml.ents() encountered a tag that was invalid
for an entry's DTD type (eg, an "ateji" tag in an entry being
formatted for jmnedict; it's invalid because the jmnedict DTD has
no "&ateji;" entity), it would print a warning message to stderr.
This is desirable when the xml formatting is done by a program
like entrs2xml.py, but when it occurs in a web page (eg by clicking
the "jmnedict xml" link in the "entr" page when viewing a jmdict
entry with a tag like "ateji"), the warning message ends up in the
web server errors log file, e.g.:
   Entry 57518: 'Illegal tag for DTD "jmnedict": \'ateji\''

This rev fixes the problem by not printing a warning message in
fmtxml.ents() but instead saving the invalid tag(s) and returning
them to the caller who can decide whether to warn or ignore them.

The returned invalid tags are ignored by default but entrs2xml.py
was updated to generate warnings about them.

Also fixed a crash in entrs2xml.py when run with neither the --xml
or --corpus options.