(from xng): Removed parse support for edict/edict2 (cgi)
Port the changes made to views/edform.py in rev 220828-a3ce097
to the corresponding CGI script.  Without these changes the cgi
version of edform fails (unable to import of the now gone edparse
module) and we need to keep cgi operational until the flask/wsgi
implementation is fully deployed.