submit.py: prohibit editing of rejected entries
Although the entry display pages do not present an Edit button for
rejected entries it is trivially easy (eg by manipulating a url or
html) to sumbit an edit for a rejected entry.  This update rejects
such submissions in the submit.py backend code.  It currently does
so for both editors and non-editors, although that could easily be
changed.