doc: rename asciidoc files from *.asc to *.adoc
In asciidoc-2.0 the interdoc xrefs broke.  An xref like:
  xref:oper.asc#upgrading[Upgrading JMdictDB]
generates a link in the product html file to "oper.asc" rather than
to "oper.html" as it did in asciidoc-1.x.  Could not find much related
directly to the issue on internet but did find one vague mention:
  "I can tell you from the 1,000 foot view that one problem is the .asc
  file extension. We're standardizing on .adoc, and certain features only
  work if you use that file extension."
  [https://github.com/asciidoctor/asciidoctor/issues/3541#issuecomment-574393892-permalink]
Renamed all the *.asc files to *.adoc.  Will update the xrefs to match
in the next commit.